fix format + merge ogden
