+++ ikarus/finiteelements/mechanics/materials/hyperelasticity/blatzko.hh
