--- ikarus/finiteelements/mechanics/materials/hyperelasticity/modified/CMakeLists.txt
@@ -1,9 +1,0 @@
-# SPDX-FileCopyrightText: 2021-2024 The Ikarus Developers mueller@ibb.uni-stuttgart.de
-# SPDX-License-Identifier: LGPL-3.0-or-later
-
-# install headers
-install(
-  FILES blatzko.hh ogden.hh
-  DESTINATION
-    ${CMAKE_INSTALL_INCLUDEDIR}/ikarus/finiteelements/mechanics/materials/hyperelasticity/modified
-)
