+++ ikarus/finiteelements/mechanics/materials/tags.hh
@@ -37,8 +37,6 @@ MAKE_ENUM(TangentModuliTags, Material, Spatial, TwoPoint);
  * \ingroup StretchTag
  * \brief A strongly typed enum class representing the type of the the principal strechts used in hyperelastic materials
  */
-MAKE_ENUM(StretchTag, principal, deviatoric);
+MAKE_ENUM(PrincipalStretchTag, total, deviatoric);
 
 } // namespace Ikarus
-
-
